Move AppleSignInController and AppleSignInControllerAuthAdapter to production
